refactor: simplify audit helpers and consolidate test stubs
Hoists sanitize_diff_panel regex patterns into static locals so they
are built once instead of per call; collapses render_pagination_links
parameter sprawl onto the FilterParams shape; trims long docblocks to
their WHY rationale; moves duplicated WP function stubs into the
shared test bootstrap.
