+++ tests/ApiHandlerExtractTest.php
@@ -91,8 +91,8 @@ final class ApiHandlerExtractTest extends TestCase {
 
 		$context = $this->logger->errors[0]['context'];
 		self::assertSame( 'chatcmpl-y', $context['id'] );
-			// Fall back to the configured model when the payload omits it.
-			self::assertSame( 'gpt-5.5', $context['model'] );
+		// Fall back to the configured model when the payload omits it.
+		self::assertSame( 'gpt-5.5', $context['model'] );
 	}
 
 	public function test_responses_output_text_helper_returns_trimmed_summary(): void {
@@ -166,8 +166,8 @@ final class ApiHandlerExtractTest extends TestCase {
 		self::assertCount( 1, $this->logger->errors );
 		$context = $this->logger->errors[0]['context'];
 		self::assertSame( 'Responses', $context['api_type'] );
-			// Fall back to the configured model when the payload omits it.
-			self::assertSame( 'gpt-5.5', $context['model'] );
+		// Fall back to the configured model when the payload omits it.
+		self::assertSame( 'gpt-5.5', $context['model'] );
 	}
 
 	public function test_responses_output_without_message_items_returns_invalid_response_error(): void {
