+++ tests/HelperApiKeyTest.php
@@ -26,8 +26,8 @@ final class HelperApiKeyTest extends TestCase {
 			'project key is accepted'         => array( 'sk-proj-ABCDEF1234567890', true ),
 			'service account key is accepted' => array( 'sk-svcacct-ABCDEF1234567890', true ),
 			'admin key is accepted'           => array( 'sk-admin-ABCDEF1234567890', true ),
-				// Keep sk-* permissive for future key flavours.
-				'unknown future flavour passes'   => array( 'sk-future-flavor-xyz', true ),
+			// Keep sk-* permissive for future key flavours.
+			'unknown future flavour passes'   => array( 'sk-future-flavor-xyz', true ),
 			'no prefix is rejected'           => array( 'totally-not-a-key', false ),
 			'uppercase prefix is rejected'    => array( 'SK-AAAAAAAAAAAA', false ),
 			'partial prefix is rejected'      => array( 's-AAAAAAAAAAAA', false ),
